Split off util functions to separate file
