+++ src/extension.ts
@@ -5,10 +5,11 @@ import * as path from 'path';
 import * as fs from 'fs-extra';
 import { platform } from 'os';
 import * as extract from 'extract-zip';
-import * as msg from './qore_message';
 import { t, addLocale, useLocale } from 'ttag';
 import * as gettext_parser from 'gettext-parser';
 import { WorkspaceFolder, DebugConfiguration, ProviderResult, CancellationToken } from 'vscode';
+import * as msg from './qore_message';
+import { compareVersion, downloadFile, findScript, openInBrowser } from './utils';
 
 export interface QoreTextDocument {
     uri: string;
@@ -82,118 +83,6 @@ function setLocale() {
     }
     else {
         msg.log(t`UsingLocaleSettings ${locale}`);
-    }
-}
-
-function compareVersion(v1, v2) {
-    if ((typeof v1 !== 'string') || (typeof v2 !== 'string')) {
-        return undefined;
-    }
-    v1 = v1.split('.');
-    v2 = v2.split('.');
-    const k = Math.min(v1.length, v2.length);
-    for (let i = 0; i < k; ++ i) {
-        v1[i] = parseInt(v1[i], 10);
-        v2[i] = parseInt(v2[i], 10);
-        if (v1[i] > v2[i]) {
-            return 1;
-        }
-        if (v1[i] < v2[i]) {
-            return -1;
-        }
-    }
-    return v1.length == v2.length ? 0: (v1.length < v2.length ? -1 : 1);
-}
-
-function findQoreScript(context: vscode.ExtensionContext, scriptName: string): string {
-    if (path.isAbsolute(scriptName)) {
-        return scriptName;
-    }
-    // try extension directory
-    let s = path.join(context.extensionPath, scriptName);
-    if (fs.existsSync(s)) {
-        return s;
-    }
-    // try PATH environment variable
-    const pathArr = (process.env.PATH || "").split(path.delimiter);
-    for (let p of pathArr) {
-        s = path.join(p, scriptName);
-        if (fs.existsSync(s)) {
-            return s;
-        }
-    }
-    return scriptName;
-}
-
-//! open an URL in the browser
-function openInBrowser(url: string) {
-    // open it in external tool - system should find appropriate handlers for schemas
-    // vscode.commands.executeCommand('vscode.open', vscode.Uri.parse(url));
-    let executable: string;
-    switch (process.platform) {
-        case 'aix':
-        case 'freebsd':
-        case 'linux':
-        case 'openbsd':
-        case 'sunos':
-            executable = 'xdg-open';
-            break;
-        case 'darwin':
-            executable = 'open';
-            break;
-        case 'win32':
-            executable = 'start';
-            break;
-        default:
-            executable = '';
-    }
-    const command: string = executable + ' ' + url;
-    try {
-        child_process.execSync(command);
-    }
-    catch (e) {
-        console.log(e);
-    }
-}
-
-function downloadFile(uri: string, dest: string, onSuccess, onError) {
-    //console.log("downloading file: " + uri);
-    //console.log("destination: " + dest);
-
-    const url = require('url');
-    const https = require('https');
-    const http = require('http');
-
-    let protocol = url.parse(uri).protocol.slice(0, -1);
-    let file = fs.createWriteStream(dest);
-
-    let localOnError = function(error) {
-        console.log("error: " + error);
-        fs.unlink(dest, err => { onError(err); }); // Delete the file async. (But we don't check the result)
-        onError(error);
-    };
-
-    let dloadFunc = function(response) {
-        if (response.statusCode >= 200 && response.statusCode < 300) {
-            file.on('error', localOnError);
-            file.on('finish', function() {
-                //console.log("Download success");
-                file.end();  // close() is async, call cb after close completes.
-                onSuccess();
-            });
-            response.pipe(file);
-        } else if (response.headers.location) { // handle redirect
-            downloadFile(response.headers.location, dest, onSuccess, onError);
-        } else {
-            localOnError(new Error("Server error: " + response));
-        }
-    };
-
-    if (protocol == "https") {
-        https.get(uri, dloadFunc).on('error', localOnError);
-    }
-    else if (protocol == "http") {
-        http.get(uri, dloadFunc).on('error', localOnError);
     }
 }
 
@@ -271,7 +160,7 @@ function isQoreVscodePkgInstalled(extensionPath: string): boolean {
 
 //! get arguments for starting QLS
 function getServerArgs(context: vscode.ExtensionContext): string[] {
-    return [findQoreScript(context, path.join("qls", "qls.q"))];
+    return [findScript(context.extensionPath, path.join("qls", "qls.q"))];
 }
 
 //! options to control the language client
@@ -848,7 +737,7 @@ export async function activate(context: vscode.ExtensionContext) {
     doQLSLaunch(context, useQLS, false);
 
     // modify debugAdapter to "/qvscdbg-test" and executable to "bash" just in case the adapter silently won't start and check command it log
-    debugAdapter = findQoreScript(context, vscode.workspace.getConfiguration("qore").get("debugAdapter") || "qdbg-vsc-adapter");
+    debugAdapter = findScript(context.extensionPath, vscode.workspace.getConfiguration("qore").get("debugAdapter") || "qdbg-vsc-adapter");
     let debuggerOk = checkDebuggerOk(qoreExecutable, debugAdapter);
     if (!debuggerOk) {
         msg.error(t`DebugAdapterNotFound '${debugAdapter}'`);
